wellbot/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
twilio==8.10.0
APScheduler==3.10.4
mysql-connector-python==8.2.0
pytesseract==0.3.10
Pillow>=10.1.0
werkzeug==3.0.1
marshmallow==3.20.1
google-auth==2.23.4
google-generativeai==0.8.4
requests==2.32.3
